+++ components/2FABundle/bundle/Resources/config/services.yaml
@@ -23,3 +23,4 @@ services:
         tags: ['controller.service_arguments']
 
     Novactive\Bundle\eZ2FABundle\Core\QRCodeGenerator: ~
+    Novactive\Bundle\eZ2FABundle\Core\UserRepository: ~
